+++ include/tao/pegtl/contrib/internal/endian.hpp
@@ -1,13 +1,13 @@
 // Copyright (c) 2017-2020 Dr. Colin Hirsch and Daniel Frey
 // Please see LICENSE for license or visit https://github.com/taocpp/PEGTL/
 
-#ifndef TAO_PEGTL_INTERNAL_ENDIAN_HPP
-#define TAO_PEGTL_INTERNAL_ENDIAN_HPP
+#ifndef TAO_PEGTL_CONTRIB_INTERNAL_ENDIAN_HPP
+#define TAO_PEGTL_CONTRIB_INTERNAL_ENDIAN_HPP
 
 #include <cstdint>
 #include <cstring>
 
-#include "../config.hpp"
+#include "../../config.hpp"
 
 #if defined( _WIN32 ) && !defined( __MINGW32__ ) && !defined( __CYGWIN__ )
 #include "endian_win.hpp"