+++ include/tao/pegtl/contrib/internal/endian_gcc.hpp
@@ -1,8 +1,8 @@
 // Copyright (c) 2017-2020 Dr. Colin Hirsch and Daniel Frey
 // Please see LICENSE for license or visit https://github.com/taocpp/PEGTL/
 
-#ifndef TAO_PEGTL_INTERNAL_ENDIAN_GCC_HPP
-#define TAO_PEGTL_INTERNAL_ENDIAN_GCC_HPP
+#ifndef TAO_PEGTL_CONTRIB_INTERNAL_ENDIAN_GCC_HPP
+#define TAO_PEGTL_CONTRIB_INTERNAL_ENDIAN_GCC_HPP
 
 #include <cstdint>
 #include <cstring>