+++ include/tao/pegtl/contrib/internal/peek_mask_uint.hpp
@@ -1,15 +1,15 @@
 // Copyright (c) 2018-2020 Dr. Colin Hirsch and Daniel Frey
 // Please see LICENSE for license or visit https://github.com/taocpp/PEGTL/
 
-#ifndef TAO_PEGTL_INTERNAL_PEEK_MASK_UINT_HPP
-#define TAO_PEGTL_INTERNAL_PEEK_MASK_UINT_HPP
+#ifndef TAO_PEGTL_CONTRIB_INTERNAL_PEEK_MASK_UINT_HPP
+#define TAO_PEGTL_CONTRIB_INTERNAL_PEEK_MASK_UINT_HPP
 
 #include <cstddef>
 #include <cstdint>
 
-#include "../config.hpp"
+#include "../../config.hpp"
+#include "../../internal/input_pair.hpp"
 
-#include "input_pair.hpp"
 #include "read_uint.hpp"
 
 namespace TAO_PEGTL_NAMESPACE::internal