+++ include/tao/pegtl/contrib/internal/peek_mask_uint8.hpp
@@ -1,19 +1,19 @@
 // Copyright (c) 2018-2020 Dr. Colin Hirsch and Daniel Frey
 // Please see LICENSE for license or visit https://github.com/taocpp/PEGTL/
 
-#ifndef TAO_PEGTL_INTERNAL_PEEK_UINT8_HPP
-#define TAO_PEGTL_INTERNAL_PEEK_UINT8_HPP
+#ifndef TAO_PEGTL_CONTRIB_INTERNAL_PEEK_MASK_UINT8_HPP
+#define TAO_PEGTL_CONTRIB_INTERNAL_PEEK_MASK_UINT8_HPP
 
 #include <cstddef>
 #include <cstdint>
 
-#include "../config.hpp"
-
-#include "input_pair.hpp"
+#include "../../config.hpp"
+#include "../../internal/input_pair.hpp"
 
 namespace TAO_PEGTL_NAMESPACE::internal
 {
-   struct peek_uint8
+   template< std::uint8_t M >
+   struct peek_mask_uint8
    {
       using data_t = std::uint8_t;
       using pair_t = input_pair< std::uint8_t >;
@@ -24,7 +24,7 @@ namespace TAO_PEGTL_NAMESPACE::internal
       template< typename ParseInput >
       [[nodiscard]] static pair_t peek( const ParseInput& in, const std::size_t /*unused*/ = 1 ) noexcept
       {
-         return { in.peek_uint8(), 1 };
+         return { std::uint8_t( in.peek_uint8() & M ), 1 };
       }
    };