+++ include/tao/pegtl/contrib/uint8.hpp
@@ -1,15 +1,15 @@
 // Copyright (c) 2018-2020 Dr. Colin Hirsch and Daniel Frey
 // Please see LICENSE for license or visit https://github.com/taocpp/PEGTL/
 
-#ifndef TAO_PEGTL_UINT8_HPP
-#define TAO_PEGTL_UINT8_HPP
+#ifndef TAO_PEGTL_CONTRIB_UINT8_HPP
+#define TAO_PEGTL_CONTRIB_UINT8_HPP
 
-#include "config.hpp"
+#include "../config.hpp"
+#include "../internal/result_on_found.hpp"
+#include "../internal/rules.hpp"
 
 #include "internal/peek_mask_uint8.hpp"
 #include "internal/peek_uint8.hpp"
-#include "internal/result_on_found.hpp"
-#include "internal/rules.hpp"
 
 namespace TAO_PEGTL_NAMESPACE::uint8
 {